# api-documentation/swagger/api.yaml
...
openapi: 3.0.3
info:
  title: Warehouse api
  description: Warehouse distribution api
  contact:
    name: Lum Berisha
    email: lum.berisha@frakton.com
  version: '1.0'
servers:
  - url: http://localhost:8080
    # description: Inferred Url
...
paths:
...
  /api/employee/warehouse/{id}:
    get:
      tags:
        - Employee
      summary: getEmployeesByWarehouse
      operationId: getEmployeesByWarehouseUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/EmployeeResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/order/warehouse/{warehouseId}:
    get:
      tags:
        - Order
      summary: getOrdersByRole
      operationId: getOrdersByRoleUsingGET
      parameters:
        - name: authenticated
          in: query
          required: false
          style: form
          schema:
            type: boolean
        # - name: authorities[0].authority
        #   in: query
        #   required: false
        #   style: form
        #   schema:
        #     type: string
        # - name: credentials
        #   in: query
        #   required: false
        #   style: form
        #   schema:
        #     type: object
        # - name: details
        #   in: query
        #   required: false
        #   style: form
        #   schema:
        #     type: object
        # - name: principal
        #   in: query
        #   required: false
        #   style: form
        #   schema:
        #     type: object
        - name: warehouseId
          in: path
          description: warehouseId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/product:
    get:
      tags:
        - Product
      summary: getProducts
      operationId: getProductsUsingGET
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ProductResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Product
      summary: createProduct
      operationId: createProductUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDto'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ProductResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/product/warehouse/{warehouseId}:
    get:
      tags:
        - Product
      summary: getProductsByWarehouse
      operationId: getProductsByWarehouseUsingGET
      parameters:
        - name: warehouseId
          in: path
          description: warehouseId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ProductResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/warehouse/products/{warehouseId}:
    post:
      tags:
        - Warehouse
      summary: setListOfProducts
      operationId: setListOfProductsUsingPOST
      parameters:
        - name: warehouseId
          in: path
          description: warehouseId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListOfProductsRequestDto'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/WarehouseResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
components:
  schemas:
...
    ListOfProductsRequestDto:
      title: ListOfProductsRequestDto
      type: object
      properties:
        listOfProducts:
          type: array
          items:
            type: string
            format: uuid
...
    ProductRequestDto:
      title: ProductRequestDto
      required:
        - name
      type: object
      properties:
        description:
          type: string
        expirationDate:
          type: string
          format: date
        manufacturingDate:
          type: string
          format: date
        name:
          type: string
        price:
          type: number
          format: double
...
    EmployeeResponseDto:
      title: EmployeeResponseDto
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponseDto'
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        address:
          $ref: '#/components/schemas/AddressResponseDto'
        vehicle:
          $ref: '#/components/schemas/VehicleResponseDto'
    VehicleResponseDto:
      title: VehicleResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        plateNumber:
          type: string
        personCapacity:
          type: integer
          format: int32
        vehicleType:
          type: string
          enum:
            - PICKUP
            - VAN
            - TRACK
    AddressResponseDto:
      title: AddressResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        region:
          type: string
          enum:
            - GJAKOVE
            - PEJE
            - PRISHTINE
            - PRIZREN
        street:
          type: string
        city:
          type: string
        postalCode:
          type: integer
          format: int32
    OrderResponseDto:
      title: OrderResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - CREATED
            - DELIVERED
            - IN_PROGRESS
            - WAITING_FOR_DELIVERY
        orderDate:
          type: string
          format: date
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponseDto'
    OrderItemResponseDto:
      title: OrderItemResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        product:
          $ref: '#/components/schemas/ProductResponseDto'
        quantity:
          type: integer
          format: int32
    ProductResponseDto:
      title: ProductResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
        description:
          type: string
        expirationDate:
          type: string
          format: date
        manufacturingDate:
          type: string
          format: date
    UserResponseDto:
      title: UserResponseDto
      properties:
        id:
          type: string
          format: uuid
        userName:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - CLIENT
            - OFFICE_WORKER
            - DELIVERY
            - WAREHOUSE_WORKER
    WarehouseResponseDto:
      title: WarehouseResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          format: string
        address:
          $ref: '#/components/schemas/AddressResponseDto'
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header